added dynamic hiding of backstage tasks based on user status
If the current user is not logged in, the user and space tabs should not
be visible. Otherwise, the login tab should be hidden.

The current implementation is less than ideal because it requires the
username to be known when backstage.show is invoked. However, that value
is determined asynchronously (triggered by TiddlyWebConfig), so the
backstage might be opened before the correct username is known.