Start support for controlling the view of bags and recipes based on the current space.
This is a fairly significant change, requiring:

* tiddlyweb 1.1.x, for filters on bag and recipe lists
* A new version of tiddlywebplugins.devstore that works
  with tidldyweb 1.1.x
* tidldywebplugins.mselect to handle the filtering

While exploring options for how to control views I was thinking
of two options:

* either a piece of middleware that interpreted the URI and
  then sent 404s where they did not match a set of requires
* or a wrapper on the store which would control what could
  be seen

Neither of these were satisfactory as they made some set of URLs
to complex to manage, for example the piece of middleware was no
go for /bags and /recipes urls

FND had the excellent insight that what was needed here was a
combo of filters and the middleware. In the middleware we look
at the uris involved and do one of the following if /bags or /recipes
are in the URI:

* if we're at the home home, just carry on like normal
* create some filters to filter the bag or recipe list, and push
  them into the environment for later processing
* if 'tiddlers' in the uri, compare the bag or recipe against the
  valid bag or recipes of the space

What's not handled:

* An explict request to /bags/some_bag or /recipes/some_recipe
  will still be processed, but if links from that are followed
  (to see tiddlers) they will fail or succeed according to the
  rules above. This is not handled simply because of a lack of
  time while working on the code and it not being critical. If
  it is needed, it will be added in the same block of code.

Better tests are needed to cover all the vectors. This was
test driven discovery as much as development as the spec is
loose.

The MiddleWare is in the root_handler module because private
code it is needed is in there. Either the module should be
renamed or bits of code should be refactored to appropriate places.