Fix a bug adding a member to a space.
The code was adding the new member to the read constraint
on the policies of the public bags and recipes. The fix
checks to see if the constraint is read and if its value is [].
If it is, then the add is skipped.

This works because the definition of public is:

   read: []

meaning anyone can read.