converted front page to imitate a space
Using the existing paradigms reduces cognitive load and allows the use
of existing functionality for managing this component (e.g. adding
editors is just a matter of adding members).

This essentially meant renaming the frontpage bag and recipe to
frontpage_public and creating dummy entities for the private subspace.

The imitation is not perfect (e.g. policies use roles, entities have a
description) which might lead to issues down the line.