Add get_identities handler.
At /users/{username}/identities get a list of those
identities which a user has associated with themselves
via the MAPUSER facility.

The implementation simple loads the tidlers from the mapuser
bag and filters them on the mapped_user field. The current
user must match the {username} or be an ADMIN.

This route is experimental, we're not sure if it will continue
living here or not.