Correct the code that guards against duplication when doing mutual subscription.
The main issue here is that Recipe.get_recipe() returns a list of lists
whereas recipes are often/usually managed as lists of tuples. Since
get_recipe is used during the subscription process we need to check
for a list of [bag, filter] not a tuple of (bag, filter) to see if the
space has already been subscribed.