changed cloneTiddler command to trigger edit mode
Instead of creating a new tiddler on the server, we only operate on the
local tiddler until the user decides to save.

Note that to avoid complexity, pressing cancel in edit mode will
currently not restore the external tiddler.

Also, cloneTiddler does not currently replace editTiddler, allowing
users to view the tiddler source without cloning.