replaced editTiddler command with cloneTiddler for external tiddlers
For now, cloneTiddler's button label (though not tooltip) imitates
editTiddler's to make it appear as the same operation to users. This
might be reverted later on.

Note that this imitiation of editTiddler does not extend to special
attributes, namely default toolbar action ("+"), requiring users to more
explicitly select this operation than when editing existing tiddlers.