Add rudimentary support for server-side unsubscribe.
Reading the comments in the test file will probably be instructive
about the complexity and caveats of how this works.

The data structure allows for one request to perform multiple
subscriptions and unsubscriptions for the same space. This is
not likely, but it makes the data structure easy to interpret.