If a login is done on the frontpage for a user that does not exist, redirect to the frontpage.
This is related to issue #8

The user remains logged in, so the frontpage can detect the situation by
the current user (form /status) not being GUEST and presenting a message
accordingly. That is:

* if the user has logged in as a non existing user, they will not be
  redirect to a space
* if the user does log in as an existing user, they do get redirected
* therefore if the user is not GUEST and we are on the frontage then
  the user must not exist in the database

Not yet closing the ticket as there are client side things to do to.

computer dieing, must send