first stab at integrating Bubbles theme to front page
Taken MartinBudden's initial work and FND's integration and tweaked the css
so that it now resembles how the front page is supposed to look.

Also changed the PageTemplate so that the Bubbles part now takes the form
of the header, leaving tiddlers (such as an about, or FAQ) to pop up
underneath.