Add a css_uri to the default configuration.
This is used by both HTMLPresenter and tiddlywebplugins.prettyerror.

It is currently set to the tiddlyweb.css at peermore.com.
IT SHOULD NOT STAY THIS.

I set it to that to have it set to something. It should probably
point to a CSS file maintained in a bag populated from the tiddlers
in src.