Added some custom scrolling behaviour
Clicking a link now scrolls the footer up near the top of the page.
This is done so that the content of the tiddler opened up is better exposed
as it scrolls the bubbles off the screen.

This has only limited use at the moment, as it is expected that in the future,
extra content/tiddlers will be added to the frontpage (such as about, help,
or FAQ tiddlers), which will make the scrolling behavious more important.