Add server_host dictionary to /status.
We need, from any tiddlyspace domain, logged in, or not logged in,
to know what the host information is for the main server. The
tiddlyweb.config server_host information is added as a dictionary
to the data provided at /status. The data looks like this:

curl http://0.0.0.0:8080/status

{"username": "GUEST",
 "challengers": [
   "cookie_form", "tiddlywebplugins.tiddlyspace.openid"
 ],
 "version": "1.1.dev3",
 "server_host": {
     "host": "tiddlyspace.org",
     "scheme": "http",
     "port": "8080"}
}

We might like to move this into tiddlywebplugins.virtualhosting
at some later date.