Removed the discoursive, social link, and added Toolbar
The discoursive, social link was to a pdf, so I've removed it, on the
basis that we should recreate it later (but to a tiddler, instead of a
pdf).

Also, Added the toolbar commands into the ViewTemplate, so that people
can play around with the TiddlyWiki-ness on the frontpage.

Added the title back into the ViewTemplate as well, in preparation for
adding a FAQ (among others).