Add support for tiddlywebplugins.whoosher-based indexing.
In the process update some requirements in setup.py.

The index replaces the default /search functionality with
one that is tied to a Whoosh index. This provides a full free text
index of the tiddlers on the system, using a default schema.

That schema can be changed if desired.

For the sake of doing following it is possible to form a query like

/search?q=ftitle:"some tiddler title" (fbag:one_public OR \
     fbag:two_public OR fbag:three_public OR ...)

To get all the public tiddlers in the followed spaces one, two, and
three that are named "some tiddler title".

NOTE: To get all the tiddlers, including those not included in
the current space, this request _must_ be made to
http://tiddlyspace.com/search, not to  http://<space>.tiddlyspace.com/.
This is because of the ControlView filter thing, which puts filters
on search and other uris to protect the pope. This can be changed if
necessary.

This is not necessarily the final syntax for how these searches can
or should be done, nor is this necessarily the final URI, this is put in
place for experimenting with the UI and data flow for following.

This implementation does not yet use tiddlywebplugins.dispatcher
but will soon.