Temporarily disable the whoosher indexing.
In the process add in a test file that never got comitted,
but with x prepended to the name to disable the test.

Doing this instead of backing out the commit because a) I want
this visible bit of history b) need to do a quick healthy deploy
to the server.