added icon identifying a story tiddler as public/private
This is just a first stab; style tweaks (size, spacing) and
optimizations to follow - along with support for avatars on external
tiddlers.

NB: The icon is displayed along with the title (view macro) - so if that
is removed or hidden, so will the icon. This is currently intentional,
based on the assumption that title and icon are a unit.