added savePublicTiddler command
This allows creating a public tiddler directly in edit mode. It is
currently limited to new tiddlers, as it's not meant to replace the
existing publishTiddlerRevision command (yet?).

There is no icon yet; it would be nice to use the saveTiddler icon in
the conventional TiddlySpace blue - perhaps also changing the regular
saveTiddler color to pink accordingly (similar to the method used on
editTiddler/cloneTiddler).