ToolbarCommands added as a separate tiddler
Previously TiddlySpaceConfig edited the shadow tiddler ToolbarCommands
to provide additional icons.
By including a tiddler called ToolbarCommands in default tiddlyspace bag
we get the same effect (user cannot delete it but can override it with
their own custom ToolbarCommands)

It also makes it easier for users to remove publish commands or revision
commands if they want to. Consider the use case where a user wants to
work purely in private.