Toolbar macro in TiddlySpace now takes extra parameters
So that users can revert back to TiddlyWiki toolbar behaviour easily
2 parameters have been added.
1: (icons:yes) allows users to turn on icons (and thus turn them off)
the default is off (a la TiddlyWiki)
2: more:popup allows users to make the more command behave as a popup
rather than a slider. Slider is default (a la TiddlyWiki)
this allows users to make use of the popup without icons
or the icons with slider. interesting stuff.

also added a missing semicolon