code cleanup + removed down glyphs from backstage
the glyphs are unnecessary and increase visual clutter to backstage
buttons.
added class userSiteIcon to user SiteIcons to be consistent with the
spaceSiteIcon class on the space SiteIcon.
the removal of glyphs required updates to the stylesheet to retain appearance
updated comments to make code more readable in terms of where overrides
are occuring on buttons