provided new toolbar icons with css effects
added simpler less visual toolbar icons
added css hover effects
experimenting with green/red/gray backgrounds
green = safe
red = beware when using this (eg. delete)
gray = harmless

also removed redundant css rule on cloneTiddler
it is expected that with new ui, the icon showing a tiddler is external
is enough to show the user something different is happening here.
best if icon is the same - we dont want to force yet another icon on the
user especially when effect is the same as another icon