Disabling prettyerror for the time being.
Prettyerror tends to leave the logs and tracebacks a bit cloudy,
and right now no content has been written for prettyerror, so all
we get by using it are errors that are formatted in a less readable
fashion.

The plugin is just commented out, all the infrastructure is still there,
waiting for the time when some real content is available.