New UI css.
The new UI has been defined in several new tiddlers.
There is no longer a need for StyleSheetToolbar which has replaced with a
more useful StyleSheetTiddlySpace.

StyleSheetTiddlySpace has been edited to remove legacy style definitions and
add references to new stylesheets.
StyleSheetBackstagePanel is included in StyleSheetTiddlySpace as these css
definitions are not essential for the functioning of the backstage bar and it
allows users to customise it later.