enabled cloneTiddler command for non-existing and shadow tiddlers
This ensures that story tiddlers which do not exist in the store are
saved to the current space when editing. In particular, this fixes
undesirable field inheritance:
http://github.com/tiddlyspace/tiddlyspace/issues#issue/148/comment/323792

This also fixes an issue with the savePublicTiddler command expecting a
workspace field to be present, which was not the case for shadow
tiddlers.