altered handling of view..SiteIcon for shadows. changes to concertina behaviour.
The changes to the view macro are so that in the case of shadow tiddlers or
tiddlers with no modifier the modifier is shown as Unknown and an image
UnknownModifierIcon is shown if available (this allows us to handle this
situation visually) This required moving of code around (less scary than
the diff looks)

In the tiddler originMacro the code has been refactored to allow the concertina
to be toggled on and off by different macros (for instance the following plugin will
use the same concertina for information). Also made minor cleanup to code.