refactored toolbar to look for icons with name without .svg suffix
The image macro can now handle binary images as well as svgs.
Therefore the icons with the suffix .svg suggest a restriction to svg
images.
Now it looks for a tiddler with the same name as the command name (no .svg suffix)
If a tiddler doesn't exist it tries the .svg suffix for backwards compatibility
reasons.