providing new commands for TiddlySpace publishing
The new commands are
* publishTiddler (copies a tiddler in a private bag into a public or vice versa)
* changeToPrivate (moves a tiddler in public to private)
* changeToPublic (moves a tiddler in private to public)
* deletePublicTiddler (delete any public versions of a tiddler)
* deletePrivateTiddler (delete any private)
* saveDraft (allows you to create a private copy of a public tiddler with
a [draft{num}] prefix which has a special custom field to allow plugins to publish
that draft to overwrite the original public).

note some of these commands are incomplete or may require a refactor - for instance
publishTiddler maybe should be separated into copyToPrivate and copyToPublic.

These commands are however not expected to be exposed in the conventional way and
will be used later by TiddlySpaceTiddlerIcons.