added new plugin ToggleTiddlerPrivacy
provides radio button to toggle the state of a new tiddler
(either new as in has no space OR new as in exists in an external space
and is being copied across).
Note: in the external case this currently does not work. Have spoken to FND
about dealing with this and have suggested use of the server.bag field.
The logic is in there so should just work if this situation is resolved.