extended cloneTiddler to remove excludeLists from pseudo-shadow tiddlers
Some tiddlers in the tiddlyspace bag take the role of shadow tiddlers.
Cloning these should not result in them being tagged with excludeLists
by default.

NB: This introduces a minor bug, as the tag will not be restored (in the
current session) of the clone operation is canceled. However, that seems
like an edge case not worth the complexity of fixing (yet).