rewrite of TiddlerIconsMacro
The original plugin had to be refactored to support
* showing icons for shadow tiddlers / missing tiddlers
(you can now use a shadowIcon to represent a tiddler being a shadow tiddler)
* unknown modifiers and the (missing) author which surfaces for new tiddlers now
make use of missingIcon if available and are displayed as unknown.
* the concertina now provides buttons that interface with commands that are
defined in the plugin TiddlySpacePublishingCommands.js - which has been pushed in another
commit that simplifies the publishing behaviour.