Correct undefined variable problem in TiddlySpaceChangePassword
Without this change a correct password change receives no success
message as the javascript errors out before the message can be
delivered.

This was discovered while attempting to address #151

However this does not seem to fix 151, it just changes the behavior.
(Now you can get it to say success when a 400 is returned, but only
after a first change that is successful)

This suggests that there is a problem with state being maintained
either in the form or the tiddlyweb.User objects. Not sure where.

I'm hoping this change pushes us ever so slightly in the direction
of getting it fixed.