Ensure that the <space>SetupFlag tiddler is saved.
Otherwise the space will attempt to initialize itself
every time it is loaded in "private mode".

The fix was determined simply by cargo-culting from elsewhere
in the plugin and then testing. The tiddler is now saved to
the recipe, which eventually means into the private bag for the
space. This appears to be what was intended by the original
code.

Addresses #185