Another style of controlling access to private bags from other spaces.
This method applies to requests under /bags/{name} or /recipes/{name}.

If the named bag is not in the recipe of the space,
    become GUEST
If the named recipe is not the public or private recipe of the space,
    become GUEST

Still not pleased with this, as it is yet more looks in the store before
we even move on.

Also, it is not clear if this works well enough. The goal is to prevent
space X, visited by user Y loading private data from user Y's stuff.

Review required.