Restrict access to more stuff.
Access to PUT/DELETE/POST /spaces /users /bags/MAPUSER can only be
done from http://tiddlyspace.com. This means that henceforth (on
this branch) a password cannot be changed from a space. This will
require some iframe handling in the backstage.

Note that the MAPSPACE bag is not included where the
MAPUSER bag is because it has a validator which sets the current
space as the thing which is mapped.