fixed ETag-related bug in TiddlySpacePublishingCommands copyTiddler
Recent enhancements to the adaptor led to the ETag taking precedence
over the revision number, thus ignoring the revision's "false" flag if a
cached ETag was present.
(This might be considered a regression in TiddlyWebAdaptor.)

This resulted in a 412 for the copyTiddler PUT and (inexplicably) a
subsequent DELETE (without an ETag) of the original tiddler - i.e. data
loss!

NB: This only addresses the most immediate issue - there are more issues
with the code here, for example:
* failed PUTs should prevent the subsequent DELETE from being triggered
  in the first place
* PUT and DELETE requests use recipes rather than bags

In conclusion, NEEDS MOAR TESTS!