made cloneTiddler a default command in the tiddler toolbar
This allows double-clicking to clone the respective tiddler. While that
reduces the explicitness of the interaction, making cloning potentially
less of a conscious decision, it was deemed more important that the
behavior be consistent with editTiddler, as the difference is too subtle
otherwise.

Note that we now have *two* default commands in the view-mode toolbar -
this works only because editTiddler and cloneTiddler are mutually
exclusive.