separated the add space form from the list space form
now you have to specify a parameter to show the add space form
or the list of spaces

Tests have been added to accompany this change.
Fixtures were updated to include getTiddlerText method.

Note this addresses http://github.com/tiddlyspace/tiddlyspace/issues#issue/198
- if a user is wanting to use the list of spaces then it is acceptable for them
to wait for the list to load.
- if a user wants to add a space, previously they would have had to wait for the list
to load. With long lists of spaces - the user would be forced to scroll to the form after
waiting for it to load. This is not a good UI.
This change puts the add space form above to make this easier on the user.