Monkeypatching the tiddlyweb.getUserInfo function
This is to modify the user.anon calculation that getUserInfo does to
make it so that only users who have signed up for an account, ie not
those who sign in via openid without previously signing up are reported
as not anonymous.

Previously, user.anon was set to true if username == "GUEST" only. In
the case of TiddlySpace, we also need to set it to true if the user
logged in isn't specifically listed in the store.

Note that this still allows the user to open up the floorboards and
dive into tiddlyweb core if they so desire, but from a UI point of
view effectively stops them from messing things up accidentally. I
believe this is the behaviour we should be aiming for.