updated user tab to make use of Logout macro.
There was a problem with the Logout code which was causing the logout
button not to render. Refactored to fix this.
Note the text here previously has been removed as it was redundant not
saying anything new - you can see username in the button heading and
the tab names show everything else.