New icons for default user/default space icons
Rather importantly this change introduces a new concept that users should have
a different default SiteIcon to that of spaces that are not associated with user
accounts.
All spaces now will be created with a default space icon
*unless* the username is the same as the current space name when it will copy the
stick man in defaultUserIcon.

There is no need to update existing users yet as this could be problematic without an
expensive checking base64 of the old SiteIcon and replacing with the new
and there is little to gain so this has been omitted and the change will just apply to
newly registered users.

Updated the login button on the backstage to use user icon.