added publishTiddler to default ToolbarCommands
If a user hacks their ViewTemplate to remove the public/private button
they then get into a situation where they cannot publish. Although a user
can discover this command and add it to their toolbar it seems a bit mean
to force them to discover it.

By placing it in toolbar commands it gives the user two methods of publishing
and allows them to decide which they prefer.

addresses http://github.com/TiddlySpace/tiddlyspace/issues#issue/244
Note previously this was discussed in
http://github.com/TiddlySpace/tiddlyspace/issues/issue/196
and removed due to "Redundant UI controls are bad because they lead to confusion."
but in light of the fact ViewTemplates are often edited and themes may not want
to use the public/private button it seems worthy of inclusion.