Implement a local store module used for overriding or extending the store.
With speed as a goal, it is useful to be able to jump right into the
storage system and pull out some data. /spaces?mine=1 is called a
rather lot and can be slow when there are a large number of spaces.
Using a direct query of the database makes it possible to not iterate
over all the recipes.

Expect more of these.