added special handling for shadow and missing icons.
Now the label saying missing only shows if a missingIcon is present. This allows a user
to make use of it (but if the icon is not present unknown modifier is rather confusing -
the absence of the image alone should hint that the modifier is unknown).

Shadow tiddlers can have their own icon shadowIcon however if this is not present it should
be treated as a core tiddler - just like tiddlers in the bag tiddlyspace and system are.