core bags no longer highlighted when excluding a space
only highlighting non core bags now.
Changed message for recursive inclusions as the previous message is now incorrect
(spaces that have the system-images bag in their recipe are including it even though it
is core and a default so the old message might mislead the user into thinking it wasn't included).
should->can to stress the fact it is optional and the user should probably look at what the services
the space provides are.
on the last exclusion, the annotation message now disappears (didn't before)
also..
increased version number
provided tiddlyspace variable for commonly used config.extensions.tiddlyspace