updated toolbar to fix issue with less/more
now if a user is using the default tiddlywiki slider syntax it should
work as normal.
If a user is using more as a popup it will treat < symbol as the end
of a menu.

There is still work that needs to be done here to allow for more than
one more menu. Adjusted code so it doesn't break anything for incompatible
menus which use > and < more than once.