new tiddlers now default to public.
Added checkbox to options panel in backstage that allows you to toggle this
setting.

Note if a user is using the setPrivacy macro they can enforce the privacy
setting by using the parameter defaultValue. If no defaultValue is given
then the config option (which is a user setting) is used.