The public, private and draft lists no longer handle edge cases
Simplified the lists so that they no longer do any ajax calls. This
is because the ajax calls were previously being done to handle the old
publishing mechanism. As this is now an edge case, we have decided not
to support it any more.

While the current three types you can list could now be put into the
core list macro, they are currently still in this one, with the aim that
TiddlySpace specific list stuff will join them. As most TiddlySpace
specific list stuff (eg spaces, members, etc) requires ajax calls, it makes
sense that this macro still exists.