updated test suite
predefined a function confirm which when run will create a confirmation
popup which when yes is clicked will run handler.

Defined a mock Popup object in setup.
renamed concertinaCommands to iconCommands in preparation of concertina
removal. Removed tests relating to privateAndPublic / privateNotPublic which
do not exist in this simplified state.