code cleanup
the feedback given by the publish is no longer relevant so discarded
to simplify the code (there is no containing element to publish a status
message in).

Added classes for styling purposes - button class now has button

fixed a bug(?) in firefox which was setting the event target to an
svg node which seems incompatible with popup. Now it finds the container
and uses that at the element to create the popup from.