added styling rules for popup.
after discussions with ben have made attempts to better present following
and privacy confirmation popups. I don't see this as "amazing" but I think
it is an improvement. Please bear in mind we proposed to use the lattice theme
as the future default theme so this is only a temporary measure.