allow disabling of certain tabs depending on user permissions
this change provides an interface for interacting with tiddlyspace and disabling
tabs that are not relevant to a given user if they are not logged in or are not
a member of a space.
A plugin must register a tab as disabled and then a hijacked tabs macro will filter
these out from parameters before rendering.

Tests are included with this commit and TiddlySpaceBackstage now registers the
private, drafts, members, file import and plugins tabs for anyone who is not a
member or not logged in.