renamed publish command, updated toolbar commands, refactored code
the publish command was confusing. publishing is usually associating with publishing
draft content. Make public/Make private is much more explicit.

changed to the changeToPublic command to be the same as publishTiddler. Instead of removing
publishTiddler completely and breaking spaces making use of it, it has been retained.

However toolbar commands updated to make use of changeToPrivate changeToPublic which allows
toggling of a tiddler state via the more menu rather than the icon which is not guaranteed
to be there. Dropped the use of show public as this is now not relevant for a default TiddlySpace
(but arguably might be useful for some use cases)