Add support for a _hash field on tiddlers.
The hash is provided by a hybrid of a change to the built in
store.py for those tiddlers which are being written and
tiddlywebplugins.hashmaker for outgoing tiddlers that do not
already have the hash.

As currently configured the hash is a sha1 digest of tiddler.text.

Any attributes can be configured for defining the hash, see the docs
for tiddlywebplugins.hashmaker