404 error message page
This is a first version of a new 404 error page.
If javascript is enabled it makes an educated guess on what the user was
trying to find.
*In cases of spaces it suggests spaces that have "similar"
names.
For instance if I visit foo.tiddlyspace.com and it doesn't exist it will suggest
food.tiddlyspace.com if this exists complete with link and SiteIcon.
it also allows me to setup the foo space if I am logged in.
* if it looks like a user was looking for a tiddler in a space
eg. foo.tiddlyspace.com/Color it suggests tiddlers in the public bag
that are similar (eg. ColorPalette).

Error page has link back to homepage at bottom for easier navigation.
I realise this is not styled but it is functionally useful and I think a
good first step.