big refactor of TiddlySpacePublishing.
Added several tests,
Under the new model it is assumed tiddlers are either in the private or public bag:
This removes the need for querying the server for public tiddlers.
Thus leads to simplifications in publisher logic.
Enabled batch publishing in the opposite direction - now works for public to private
as well as private to public.