release 0.9.30
* tests and fixes for groupBy plugin
* the lwiki serialization

The lwiki serialization is available from raw tiddlyweb urls (not
tiddlyspace) and provides a way to experiment with lazy loading
at scale. To get a lazy view of a public space:

  http://space.tiddlyspace.com/recipes/space_public/tiddlers.lwiki

(private as you would expect)

There's no expectation at this time that the lwiki serialization will
stay around or become the default, this is merely to learn and then
build better things later.