added plumbing to allow show and hide buttons to show in browsers without svg support
the image macro takes a altImage parameter in the case of svgs to degrade to buttons.
This is a consistency problem. If docs say click on the close button in the top right,
this can be confusing in the current setup as a key part of the ui is not consistent
across browsers. svgs however look nicer (scale better) so are preferable if possible.