MarkupPreHead tiddlyinited
now MarkupPreHead is constructed via TiddlyInit to contain an atom feed uri
as well as a favicon. The favicon points to recipe rather than bag url to allow
easier overriding.

Note the MarkupPreHead tiddler is not overwritten if a user already has their own
public version of it and the MarkupPreHead tiddler is retained in the tiddlyspace bag
so that on first load the default TiddlyWiki incorrect atom feed uri is not displayed.