Made the tiddlyspace icon in backstage clickable.
This now opens a popup menu currently containing a link back to the user's
homespace, the version number and a link to the several spaces defined in backstage tiddler.

message added to config.messages to display status of user in that space for use
with message macro.

This also required the addition of a homeLink macro which links to the users
homespace. it doesn't show when the user is already there.