added server side support for an archive bag
the archive bag will be used to store old content so that people can recover
"lost" revisions, etc.

It is currently added with the same permissions as the private bag, but does
not appear in any recipe.

This code only deals with the server side implementation, etra code will be
needed on the client side to make this do anything.

Note also that deploying will require an extra step. The creation of an archive
bag for each current space.