avatars now to do not link to a space that you are in
If you are in the space foo and see an avatar for foo previously this was a
clickable icon. This change (with tests) disables the link when this is the
case as there is no reason why you would want to reload your page. This has
also been confusing steve.