Added subscribe parameter to TiddlySpaceSpaces
if the parameter is present then the checkbox "Include the current space in the new space."
will be checked by default.
This is useful for spaces designed for inclusion for helping users get setup quickly
e.g familytree.tiddlyspace.com

This required updates to tests as the code now uses the paramString for parameters.
Added an extra test for dealing with the checkbox being set.