Created a function to check a nonce value to help with CSRF
The idea is that the client will PUT a nonce value (inside a tiddler) into the
private bag of a space. They will then include this value whenever POSTing
something that is (currently) susceptible to CSRF.

This provides a function that will check said value against the value that is
sent with the request.