Added reply link view type.
The replyLink view type creates a link to a tiddler in the users home space
with the same name as the attribute passed to it.
e.g. <<view title replyLink>> on a tiddler foo
for user j on tiddlyspace.com will create a link
to http://j.tiddlyspace.com#[[foo]]

note link does not show if
* tiddler from users space
* user not logged in

note intentionally if a user is including a space foo which includes a tiddler hello
when tiddler hello is opened a reply link shows which points to itself (as you can
clone).

Added a macro to the ViewTemplate however note this might not be enough as a
user not using a default TiddlySpace theme will make it impossible for users
to reply to them. Although it is important to allow a user control over how the
replyLink is presented. A suggested approach might be to override displayTiddler
to check for the existance of a class replyLink in the ViewTemplate and
render it if it doesn't exist.

Tests demonstrate the different scenarios.