Created some wsgi middleware to look for csrf
This middleware sits right after Negotiate, as it relies on
a user being present, and a content type (tiddlyweb.type).

It uses that check_csrf function to validate potential requests.

A potential request in this case is defined as a POST (all GETs are idempotent,
PUTs and DELETEs can only be done with xhr so are not potential vectors), that
has a content type of application/x-www-form-urlencoded or multipart/form-data
(other content types can only be POSTed in xhrs, so are immune).

Currently, a nonce value must be included in the form submission, in the format:

	nonce=value

this value must correspond to a tiddler (currently called "nonce") and a field
within that tiddler (also called nonce), which must match the value given.

It is expected that these names will change in the future to something more
appropriate (token for example).

Other areas for improvement include defining how the nonce is built, and the
name of the tiddler it is stored in, so that the server can verify that the
client knows what they are doing, and not just that they have the authority
to do it.