restored timeout
This is more complicated then first thought as the timeout is
needed due to the fact addCustomFields runs at the end of displayTiddler.

The setPrivacy macro not only needs to know the values passed in through
the tiddlyfields attribute but also needs to wait for custom fields to be
added before setting them.

a ticket exists:
http://trac.tiddlywiki.org/ticket/1195