Added set-cookie code, more tests and fixed time based bug
Added the code to create and set a csrf_token cookie if none present
or if the cookie is too old.

Added some tests to make sure this is happening correctly

fixed a bug whereby the time encoded in the cookie was being used to
check itself, instead of using the time _now_. Added a test for this.