ensured that RevisionsView is deactivated before entering edit mode
Previously it was possible to activate both edit mode and RevisionsView
by triggering the default action (via double-click) on a story tiddler.

cf. issue #451

In the process, also fixed hijacking of toolbar commands; no reason to
use a jQuery event if what we want is to modify the native event.

plus some minor enhancements (efficiency, documentation)